Version update and release with minor fixes

+++ app/src/main/java/org/j_keepass/LoadActivity.java
@@ -206,43 +206,61 @@ public class LoadActivity extends AppCompatActivity {
         createBtn.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                Penta<AlertDialog, MaterialButton, MaterialButton, TextInputEditText, TextInputEditText> confirmDialog = DatabaseCreateDialogUtil.getConfirmDialog(getLayoutInflater(), binding.getRoot().getContext());
 
-                confirmDialog.second.setOnClickListener(v1 -> {
-                    if (!Common.isCodecAvailable) {
-                        ToastUtil.showToast(getLayoutInflater(), v1, R.string.devInProgress);
-                    } else if (confirmDialog.fourth.getText() == null || confirmDialog.fourth.getText().toString().length() <= 0) {
-                        ToastUtil.showToast(getLayoutInflater(), v1, R.string.enterDatabaseName);
-                    } else if (confirmDialog.fifth.getText() == null || confirmDialog.fifth.getText().toString().length() <= 0) {
-                        ToastUtil.showToast(getLayoutInflater(), v1, R.string.enterPassword);
-                    } else {
-                        String dbName = confirmDialog.fourth.getText().toString();
-                        if (!dbName.endsWith("kdbx")) {
-                            dbName = dbName + ".kdbx";
-                        }
-                        File fromTo = new File(subFilesDirPath + File.separator + dbName);
-                        if (fromTo.exists()) {
-                            fromTo.delete();
+                boolean isOK = false;
+
+                if (ContextCompat.checkSelfPermission(binding.getRoot().getContext(), Manifest.permission.WRITE_EXTERNAL_STORAGE) != PackageManager.PERMISSION_GRANTED
+                        || ContextCompat.checkSelfPermission(binding.getRoot().getContext(), Manifest.permission.READ_EXTERNAL_STORAGE) != PackageManager.PERMISSION_GRANTED
+                ) {
+                    ActivityCompat.requestPermissions(LoadActivity.this, new String[]{
+                                    Manifest.permission.READ_EXTERNAL_STORAGE, Manifest.permission.WRITE_EXTERNAL_STORAGE},
+                            READ_EXTERNAL_STORAGE);
+                }
+
+                if (ContextCompat.checkSelfPermission(binding.getRoot().getContext(), Manifest.permission.READ_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED) {
+                    isOK = true;
+                } else {
+                    ToastUtil.showToast(getLayoutInflater(), v, R.string.permissionNotGranted);
+                }
+                if (isOK) {
+                    Penta<AlertDialog, MaterialButton, MaterialButton, TextInputEditText, TextInputEditText> confirmDialog = DatabaseCreateDialogUtil.getConfirmDialog(getLayoutInflater(), binding.getRoot().getContext());
+
+                    confirmDialog.second.setOnClickListener(v1 -> {
+                        if (!Common.isCodecAvailable) {
+                            ToastUtil.showToast(getLayoutInflater(), v1, R.string.devInProgress);
+                        } else if (confirmDialog.fourth.getText() == null || confirmDialog.fourth.getText().toString().length() <= 0) {
+                            ToastUtil.showToast(getLayoutInflater(), v1, R.string.enterDatabaseName);
+                        } else if (confirmDialog.fifth.getText() == null || confirmDialog.fifth.getText().toString().length() <= 0) {
+                            ToastUtil.showToast(getLayoutInflater(), v1, R.string.enterPassword);
                         } else {
-                            try {
-                                fromTo.createNewFile();
-                                kdbxFileUri = Uri.fromFile(fromTo);
-                                KdbxCreds creds = new KdbxCreds(confirmDialog.fifth.getText().toString().getBytes());
-                                Database<?, ?, ?, ?> database = getDummyDatabase();
-                                OutputStream fileOutputStream = getContentResolver().openOutputStream(kdbxFileUri, "wt");
-                                database.save(creds, fileOutputStream);
-                                confirmDialog.first.dismiss();
-                                closeKeyboard();
-                                fetchAndShowFiles();
-                                kdbxFileUri = null;
-                            } catch (Exception e) {
-                                ToastUtil.showToast(getLayoutInflater(), v1, e.getMessage());
+                            String dbName = confirmDialog.fourth.getText().toString();
+                            if (!dbName.endsWith("kdbx")) {
+                                dbName = dbName + ".kdbx";
+                            }
+                            File fromTo = new File(subFilesDirPath + File.separator + dbName);
+                            if (fromTo.exists()) {
+                                fromTo.delete();
+                            } else {
+                                try {
+                                    fromTo.createNewFile();
+                                    kdbxFileUri = Uri.fromFile(fromTo);
+                                    KdbxCreds creds = new KdbxCreds(confirmDialog.fifth.getText().toString().getBytes());
+                                    Database<?, ?, ?, ?> database = getDummyDatabase();
+                                    OutputStream fileOutputStream = getContentResolver().openOutputStream(kdbxFileUri, "wt");
+                                    database.save(creds, fileOutputStream);
+                                    confirmDialog.first.dismiss();
+                                    closeKeyboard();
+                                    fetchAndShowFiles();
+                                    kdbxFileUri = null;
+                                } catch (Exception e) {
+                                    ToastUtil.showToast(getLayoutInflater(), v1, e.getMessage());
+                                }
                             }
                         }
-                    }
 
-                });
-                DatabaseCreateDialogUtil.showDialog(confirmDialog.first);
+                    });
+                    DatabaseCreateDialogUtil.showDialog(confirmDialog.first);
+                }
             }
         });
         binding.floatInfo.setOnClickListener(new View.OnClickListener() {
